complete day 12 tasks

+++ src/main/java/day9/Task2/Triangle.java
@@ -1,11 +1,11 @@
 package day9.Task2;
 
 public class Triangle extends Figure {
-    private int a;
-    private int b;
-    private int c;
+    private double a;
+    private double b;
+    private double c;
 
-    public Triangle(int a, int b, int c, String color) {
+    public Triangle(double a, double b, double c, String color) {
         super(color);
         this.a = a;
         this.b = b;
